Add pagination

+++ gym-app/app/Http/Controllers/LockerController.php
@@ -4,9 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Gym;
 use App\Models\Locker;
-use Illuminate\Contracts\Validation\Rule;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\Rule;
 
 class LockerController extends Controller
 {
@@ -21,7 +21,7 @@ class LockerController extends Controller
             abort(403);
         }
 
-        $lockers = Locker::all();
+        $lockers = Locker::simplePaginate(12);
 
         return view('admin.lockers.index', ['lockers' => $lockers]);
     }
@@ -59,7 +59,7 @@ class LockerController extends Controller
                 'gym_id' => 'required|in:' . Gym::all()->pluck('id')->implode(','),
                 'number' => [
                     'required',
-                    'number',
+                    'numeric',
                     'min:1',
                     Rule::unique('lockers')->where(function ($query) use ($request) {
                         return $query->where('gym_id', $request->gym_id);
@@ -86,10 +86,19 @@ class LockerController extends Controller
             abort(403);
         }
 
+        if (Locker::find($id)->is_used()) {
+            abort(403);
+        }
+
         $locker = Locker::find($id);
+
+        if ($locker == null) {
+            return redirect()->route('lockers.index')->with('not-found', $id);
+        }
+
         $gyms = Gym::all();
 
-        return view('admin.lockers.create', ['locker' => $locker, 'gyms' => $gyms]);
+        return view('admin.lockers.edit', ['locker' => $locker, 'gyms' => $gyms]);
     }
 
     /**
@@ -105,20 +114,29 @@ class LockerController extends Controller
             abort(403);
         }
 
+        $locker = Locker::find($id);
+
+        if ($locker == null) {
+            abort(403);
+        }
+
+        if ($locker->is_used()) {
+            abort(403);
+        }
+
         $validated = $request->validate(
             [
                 'number' => [
-                    'number',
+                    'numeric',
                     'min:1',
                     Rule::unique('lockers')->where(function ($query) use ($request) {
                         return $query->where('gym_id', $request->gym_id);
-                    }),
+                    })->ignore($locker->id),
                 ],
                 'gender' => 'in:male,female',
             ],
         );
 
-        $locker = Locker::find($id);
         $locker->update($validated);
 
         return redirect()->route('lockers.index')->with('edit', $locker->number);
@@ -131,6 +149,10 @@ class LockerController extends Controller
         }
 
         $locker = Locker::find($id);
+
+        if ($locker->is_used()) {
+            abort(403);
+        }
 
         return view('admin.lockers.delete', ['locker' => $locker]);
     }
@@ -148,6 +170,10 @@ class LockerController extends Controller
         }
 
         $locker = Locker::find($id);
+
+        if ($locker->is_used()) {
+            abort(403);
+        }
         $locker_number = $locker->number;
         $locker->delete();
